Do not multiply the absolute error tolerance with machine epsilon

+++ dev/tools/run_tests.py
@@ -6,5 +6,6 @@ import sys
 import brian2
 
 if __name__ == '__main__':
-    if not brian2.test():  # If the test fails, exit with a non-zero error code
+    import numpy as np
+    if not brian2.test(float_dtype=np.float32):  # If the test fails, exit with a non-zero error code
         sys.exit(1)
